backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dateutil==2.8.2
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
paramiko==3.3.1
pywinrm==0.4.3
pydantic==2.5.0
email-validator==2.1.0
pyOpenSSL==23.3.0